Move BaseReorderAttributesMutation to attribute module
