+++ saleor/graphql/product/mutations/attributes.py
@@ -1,5 +1,5 @@
 from collections import defaultdict
-from typing import TYPE_CHECKING, List
+from typing import List
 
 import graphene
 from django.core.exceptions import ObjectDoesNotExist, ValidationError
@@ -10,6 +10,7 @@ from ....attribute import AttributeInputType, AttributeType, models as attribute
 from ....core.permissions import ProductTypePermissions
 from ....product import models
 from ....product.error_codes import ProductErrorCode
+from ...attribute.mutations import BaseReorderAttributesMutation
 from ...attribute.types import Attribute
 from ...core.inputs import ReorderInput
 from ...core.mutations import BaseMutation
@@ -17,11 +18,7 @@ from ...core.types.common import ProductError
 from ...core.utils import from_global_id_strict_type
 from ...core.utils.reordering import perform_reordering
 from ...product.types import ProductType
-from ...utils import resolve_global_ids_to_primary_keys
 from ..enums import ProductAttributeType
-
-if TYPE_CHECKING:
-    from django.db.models import QuerySet
 
 
 class ProductAttributeAssignInput(graphene.InputObjectType):
@@ -285,56 +282,6 @@ class ProductAttributeUnassign(BaseMutation):
         cls.save_field_values(product_type, "variant_attributes", attribute_pks)
 
         return cls(product_type=product_type)
-
-
-class BaseReorderAttributesMutation(BaseMutation):
-    class Meta:
-        abstract = True
-
-    @classmethod
-    def prepare_operations(cls, moves: ReorderInput, attributes: "QuerySet"):
-        """Prepare operations dict for reordering attributes.
-
-        Operation dict format:
-            key: attribute pk,
-            value: sort_order value - relative sorting position of the attribute
-        """
-        attribute_ids = []
-        sort_orders = []
-
-        # resolve attribute moves
-        for move_info in moves:
-            attribute_ids.append(move_info.id)
-            sort_orders.append(move_info.sort_order)
-
-        _, attr_pks = resolve_global_ids_to_primary_keys(attribute_ids, Attribute)
-        attr_pks = [int(pk) for pk in attr_pks]
-
-        attributes_m2m = attributes.filter(attribute_id__in=attr_pks)
-
-        if attributes_m2m.count() != len(attr_pks):
-            attribute_pks = attributes_m2m.values_list("attribute_id", flat=True)
-            invalid_attrs = set(attr_pks) - set(attribute_pks)
-            invalid_attr_ids = [
-                graphene.Node.to_global_id("Attribute", attr_pk)
-                for attr_pk in invalid_attrs
-            ]
-            raise ValidationError(
-                "Couldn't resolve to an attribute.",
-                params={"attributes": invalid_attr_ids},
-            )
-
-        attributes_m2m = list(attributes_m2m)
-        attributes_m2m.sort(
-            key=lambda e: attr_pks.index(e.attribute.pk)
-        )  # preserve order in pks
-
-        operations = {
-            attribute.pk: sort_order
-            for attribute, sort_order in zip(attributes_m2m, sort_orders)
-        }
-
-        return operations
 
 
 class ProductTypeReorderAttributes(BaseReorderAttributesMutation):
